Use ssize_t for record sizes.

+++ include/base_stream.h
@@ -240,7 +240,7 @@ class base_stream : public abstract_stream, public std::enable_shared_from_this<
     /// Received, waiting to be read.
     std::deque<rx_segment_t> rx_segments_;
     /// Sizes of received messages.
-    std::deque<int64_t> rx_record_sizes_;
+    std::deque<ssize_t> rx_record_sizes_;
 
     int32_t receive_buf_size_{default_rx_buffer_size};         // Recv buf size for channel control
     int32_t child_receive_buf_size_{default_rx_buffer_size};   // Recv buf for child streams
